Attack cool down
I added the code that helps the program decide when to give attack cool down. A list keeps appending 'a' until it reaches of length 3 because it enters else the list is cleared. and as there is a lock condition in else it keeps the user from attacking for 2 tries, might aswell add a small condition using random that lets the user attack even though he can't.